Build out GroupDataSource
By extracting a bit of the details of the feed provider
we're able to come up with a consistent interface for the GroupDataSource
and the EventDataSource. We now have a FeedProviderDelegate (formerly
EventFeedDelegate) protocol as well as a FeedProvider protocol (new).
These dictate the communication between the providers of the
feed and their respective delegate partners.